+++ Example/Sources/AppDelegate.swift
@@ -19,7 +19,6 @@ final class AppDelegate: UIResponder, UIApplicationDelegate {
     
     func application(_ application: UIApplication,
                      didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        DateFormatter.log = nil
         return true
     }